feat(types): centralize shared type definitions
