+++ src/renderer/src/components/HeadersEditor.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import type { RequestHeader } from '../hooks/useHeadersManager';
+import type { RequestHeader } from '../types';
 
 interface HeadersEditorProps {
   headers: RequestHeader[];
